# project-app/src/main/gen/api/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: jeff.nelwek@gmail.com
    name: Jeffrey Nelwek
  description: API zum Verwalten von Listen
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: JEFF'S REST-API
  version: v3.0
servers:
- description: Entwicklungssystem
  url: http://localhost:8080/project
paths:
  /api/user/all:
    get:
      operationId: allProjects
      responses:
        "200":
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/User'
                type: array
          description: OK
      tags:
      - user-controller
      x-accepts:
      - '*/*'
  /api/user/delete/{id}:
    delete:
      operationId: deleteUser
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                type: boolean
          description: OK
      tags:
      - user-controller
      x-accepts:
      - '*/*'
  /api/user/save:
    post:
      operationId: saveUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
          description: OK
      tags:
      - user-controller
      x-content-type: application/json
      x-accepts:
      - '*/*'
  /api/user/user/{id}:
    get:
      operationId: getUser
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User'
          description: OK
      tags:
      - user-controller
      x-accepts:
      - '*/*'
components:
  schemas:
    User:
      example:
        zipCode: zipCode
        city: city
        phone: phone
        street: street
        houseNumber: houseNumber
        name: name
        id: id
        email: email
        username: username
      properties:
        city:
          type: string
        email:
          type: string
        houseNumber:
          type: string
        id:
          type: string
        name:
          type: string
        phone:
          type: string
        street:
          type: string
        username:
          type: string
        zipCode:
          type: string
      type: object